requirements.txt
anyio==3.6.2
argcomplete==1.0.0
asd==1.3.2
async-generator==1.10
attrs==22.1.0
Automat==22.10.0
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
constantly==15.1.0
contourpy==1.0.6
cryptography==38.0.1
cssselect==1.2.0
cycler==0.11.0
distlib==0.3.6
exceptiongroup==1.0.4
fake-useragent==1.1.1
fastapi==0.85.2
filelock==3.8.0
fonttools==4.38.0
google==3.0.0
google-api-core==2.10.2
google-auth==2.14.0
google-cloud==0.34.0
google-cloud-appengine-logging==1.1.6
google-cloud-audit-log==0.2.4
google-cloud-bigquery==3.3.5
google-cloud-bigquery-storage==2.16.2
google-cloud-compute==1.6.1
google-cloud-core==2.3.2
google-cloud-datastore==2.9.0
google-cloud-logging==3.2.5
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpc-google-iam-v1==0.12.4
grpcio==1.50.0
grpcio-status==1.50.0
h11==0.14.0
hyperlink==21.0.0
idna==3.4
importlib-resources==5.10.0
incremental==22.10.0
itemadapter==0.7.0
itemloaders==1.0.6
jieba==0.42.1
jmespath==1.0.1
jsonlines==3.1.0
kiwisolver==1.4.4
lxml==4.9.1
matplotlib==3.6.2
numpy==1.23.4
outcome==1.2.0
packaging==21.3
pandas==1.5.1
parsel==1.6.0
Pillow==9.3.0
platformdirs==2.5.2
Protego==0.2.1
proto-plus==1.22.1
protobuf==4.21.9
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
PyDispatcher==2.0.6
pyOpenSSL==22.1.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.6
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
routers==0.10.1
rsa==4.9
scipy==1.9.3
Scrapy==2.7.0
selenium==4.7.2
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
starlette==0.20.4
tldextract==3.4.0
trio==0.22.0
trio-websocket==0.9.2
Twisted==22.10.0
twisted-iocpsupport==1.0.2
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
virtualenv==20.16.6
voidpp-tools==1.8.1
w3lib==2.0.1
wordcloud==1.8.2.2
wsproto==1.2.0
zipp==3.11.0
zope.interface==5.5.0
